requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
contourpy==1.1.1
cycler==0.11.0
dateparser==1.1.8
fonttools==4.42.1
frozenlist==1.4.0
idna==3.4
kiwisolver==1.4.5
matplotlib==3.8.0
multidict==6.0.4
numpy==1.26.0
packaging==23.1
pandas==2.1.1
Pillow==10.0.1
pycryptodome==3.19.0
pyparsing==3.1.1
python-binance==1.0.19
python-dateutil==2.8.2
pytz==2023.3.post1
regex==2023.8.8
requests==2.31.0
six==1.16.0
TA-Lib==0.4.28
tzdata==2023.3
tzlocal==5.0.1
ujson==5.8.0
urllib3==2.0.5
websockets==11.0.3
yarl==1.9.2
